requirements.txt
Django==3.2.8
psycopg2-binary==2.8.5
ipython==7.28.0
invoke==1.6.0
django_select2==7.9.0
django-celery-beat==2.2.1
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
celery==5.2.1
redis==4.0.2
uwsgi==2.0.20
djangorestframework-simplejwt==5.0.0
django-filter==21.1

djangorestframework==3.13.1
beautifulsoup4==4.10.0
requests==2.26.0